requirements.txt
flask==2.0.1
python-dotenv==0.19.0
twilio==7.0.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
gspread==5.0.0 